Crate aper_yew

Aper-Yew

This crate provides a Yew component that connects to an Aper server. The component takes as arguments a websocket URL, which it connects to in order to send and receive state updates. The component also takes an object implementing the [StateView] trait, which provides a function from the current state object to the [yew::Html] that should be rendered.

Note that the [StateView] you provide is not a standard Yew component. That’s because Yew components typically own either a copy of their data or a read-only reference to it. Since the state is already owned by the [StateMachineComponent] that calls the view, this model allows us to pass the state by reference whenever we need to render and thus avoid creating additional copies of the data.

This doesn’t mean [StateView]s can’t have their own state, though! StateViews can contain stateful components by embedding them in the resulting [yew::Html] just as they would in a regular Yew component.

Structs

Props
Properties for [StateMachineComponent].

StateProgramComponent
Yew Component which owns a copy of the state as well as a connection to the server, and keeps its local copy of the state in sync with the server.

UpdateInterval
A Yew component that calls the given callback at a regular interval. It is a useful way to automatically refresh a [crate::StateView], since the state view itself cannot own an [IntervalTask].

ViewContext

Enums

Msg
Represents a message this component could receive, either from the server or from an event triggered by the user.

Status
The connection status of the component, and stores the state once it is available. The component does not have a copy of the state until it has connected and received an initial copy of the server’s current state.

Traits

View