Trait aper_yew::StateView

pub trait StateView: Sized + 'static + Debug + Clone {
    type State: StateMachine + Debug;
    fn view(
        &self, 
        state: &Self::State, 
        view_context: &ViewContext<Self::State>
    ) -> Html;
}

A trait implemented by objects which can render a StateMachine into [yew::Html]. In some cases it will be useful to implement this on empty structs, such that the view is dependent entirely on the value of the StateMachine and PlayerID. In cases where this is implemented on non-empty structs, the data in the struct can be used for rendering.

Associated Types

type State: StateMachine + Debug

Defines the struct used to represent the state that this StateView renders.

Loading content...
Required methods

fn view(
&self,
state: &Self::State,
view_context: &ViewContext<Self::State>
) -> Html

Render the given state into a [yew::Html] result.

Arguments

state - The state to render.

callback - A callback which, when called, propagates a transition to the state machine. The transition is an Option, if it is None this call is a no-op.

player - Upon connecting to the websocket server, each client is assigned a PlayerID. It is passed to the view, so that the view can depend on the player who is viewing it.

Loading content...