Trait aper_yew::View

pub trait View: Clone {
    type State;
    type Callback;
    fn view(
        &self, 
        value: &Self::State, 
        context: &ViewContext<Self::Callback>
    ) -> Html;
}

Applies to a struct that can turn a value of the associated State type into yew::Html. The resulting view can emit events of type Option<Callback>.

Associated Types

type State

type Callback

Required methods

fn view(
&self,
value: &Self::State,
context: &ViewContext<Self::Callback>
) -> Html

Turn a value into HTML.